+++ R/utils.R
@@ -47,3 +47,12 @@ UseMethod("%+%")
 
 "%+%.default" <- function(x,y) paste(paste(x,collapse=","),paste(y,collapse=","),sep="")
 # we often construct warning msgs with a msg followed by several items of a vector, so %+% is for convenience
+
+require_bit64 = function() {
+  # called in fread and print when they see integer64 columns are present 
+  tt = try(requireNamespace("bit64",quietly=TRUE))
+  if (inherits(tt,"try-error"))
+    warning("Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.")
+}
+
+